fix: red packet loading
