+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketConfirmDialog.tsx
@@ -1,14 +1,27 @@
 import BigNumber from 'bignumber.js'
 import classNames from 'classnames'
-import { useEffect } from 'react'
-import { formatBalance, isNativeTokenAddress, resolveTokenLinkOnExplorer, useChainId } from '@masknet/web3-shared-evm'
+import { useCallback, useEffect, useMemo, useRef } from 'react'
+import {
+    formatBalance,
+    getChainName,
+    isNativeTokenAddress,
+    resolveTokenLinkOnExplorer,
+    TransactionStateType,
+    useAccount,
+    useChainId,
+    useNetworkType,
+    useRedPacketConstants,
+    useWeb3,
+} from '@masknet/web3-shared-evm'
 import { Grid, Link, Paper, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
 import LaunchIcon from '@mui/icons-material/Launch'
-import { FormattedBalance } from '@masknet/shared'
+import { FormattedBalance, useOpenShareTxDialog } from '@masknet/shared'
 import ActionButton from '../../../extension/options-page/DashboardComponents/ActionButton'
 import { useI18N } from '../../../utils'
-import type { RedPacketSettings } from './hooks/useCreateCallback'
+import { RedPacketSettings, useCreateCallback } from './hooks/useCreateCallback'
+import type { RedPacketJSONPayload, RedPacketRecord } from '../types'
+import { RedPacketRPC } from '../messages'
 
 const useStyles = makeStyles()((theme) => ({
     link: {
@@ -66,20 +79,121 @@ const useStyles = makeStyles()((theme) => ({
 
 export interface ConfirmRedPacketFormProps {
     onBack: () => void
-    onCreate: () => void
+    onCreated: (payload: RedPacketJSONPayload) => void
     onClose: () => void
     settings?: RedPacketSettings
 }
 
 export function RedPacketConfirmDialog(props: ConfirmRedPacketFormProps) {
     const { t } = useI18N()
-    const { onBack, settings, onCreate, onClose } = props
+    const { onBack, settings, onClose, onCreated } = props
     const { classes } = useStyles()
     const chainId = useChainId()
 
     useEffect(() => {
         if (settings?.token?.chainId !== chainId) onClose()
     }, [chainId, onClose])
+
+    // #region blocking
+    // password should remain the same rather than change each time when createState change,
+    //  otherwise password in database would be different from creating red-packet.
+    const contract_version = 4
+    const web3 = useWeb3()
+    const account = useAccount()
+    const { address: publicKey, privateKey } = useMemo(() => web3.eth.accounts.create(), [])
+    const [createSettings, createState, createCallback, resetCreateCallback] = useCreateCallback(
+        settings!,
+        contract_version,
+        publicKey,
+    )
+    const isCreating = [
+        TransactionStateType.WAIT_FOR_CONFIRMING,
+        TransactionStateType.HASH,
+        TransactionStateType.RECEIPT,
+    ].includes(createState.type)
+    const openShareTxDialog = useOpenShareTxDialog()
+    const createRedpacket = useCallback(async () => {
+        const hash = await createCallback()
+        if (!hash) return
+        openShareTxDialog({
+            hash,
+            onShare() {},
+        })
+        // reset state
+        resetCreateCallback()
+
+        // the settings is not available
+        if (!createSettings?.token) return
+
+        // TODO:
+        // early return happened
+        // we should guide user to select the red packet in the existing list
+        if (createState.type !== TransactionStateType.CONFIRMED) return
+
+        const { receipt } = createState
+        const CreationSuccess = (receipt.events?.CreationSuccess.returnValues ?? {}) as {
+            creation_time: string
+            creator: string
+            id: string
+            token_address: string
+            total: string
+        }
+        payload.current.sender = {
+            address: account,
+            name: createSettings.name,
+            message: createSettings.message,
+        }
+        payload.current.is_random = createSettings.isRandom
+        payload.current.shares = createSettings.shares
+        payload.current.password = privateKey
+        payload.current.rpid = CreationSuccess.id
+        payload.current.total = CreationSuccess.total
+        payload.current.duration = createSettings.duration
+        payload.current.creation_time = Number.parseInt(CreationSuccess.creation_time, 10) * 1000
+        payload.current.token = createSettings.token
+
+        // output the redpacket as JSON payload
+        onCreated(payload.current)
+    }, [createCallback, resetCreateCallback, createSettings, createState, openShareTxDialog, onCreated])
+    // #endregion
+
+    // assemble JSON payload
+    const payload = useRef<RedPacketJSONPayload>({
+        network: getChainName(chainId),
+    } as RedPacketJSONPayload)
+
+    const { HAPPY_RED_PACKET_ADDRESS_V4 } = useRedPacketConstants()
+    const networkType = useNetworkType()
+    useEffect(() => {
+        if (createState.type !== TransactionStateType.UNKNOWN) return
+        const contractAddress = HAPPY_RED_PACKET_ADDRESS_V4
+        if (!contractAddress) {
+            onClose()
+            return
+        }
+        payload.current.contract_address = contractAddress
+        payload.current.contract_version = contract_version
+        payload.current.network = getChainName(chainId)
+    }, [chainId, networkType, contract_version, createState])
+
+    useEffect(() => {
+        if (!createSettings?.token || createState.type === TransactionStateType.UNKNOWN) return
+
+        // storing the created red packet in DB, it helps retrieve red packet password later
+        // save to the database early, otherwise red-packet would lose when close the tx dialog or
+        //  web page before create successfully.
+        if (createState.type === TransactionStateType.HASH && createState.hash) {
+            payload.current.txid = createState.hash
+            const record: RedPacketRecord = {
+                id: createState.hash!,
+                from: '',
+                password: privateKey,
+                contract_version,
+            }
+            RedPacketRPC.discoverRedPacket(record)
+        }
+    }, [createState /* update tx dialog only if state changed */])
+    // #endregion
 
     return (
         <Grid container spacing={2} className={classNames(classes.grid, classes.gridWrapper)}>
@@ -176,12 +290,12 @@ export function RedPacketConfirmDialog(props: ConfirmRedPacketFormProps) {
             </Grid>
 
             <Grid item xs={6}>
-                <ActionButton variant="contained" size="large" fullWidth onClick={onBack}>
+                <ActionButton disabled={isCreating} variant="contained" size="large" fullWidth onClick={onBack}>
                     {t('plugin_red_packet_back')}
                 </ActionButton>
             </Grid>
             <Grid item xs={6}>
-                <ActionButton variant="contained" size="large" fullWidth onClick={onCreate}>
+                <ActionButton loading={isCreating} variant="contained" size="large" fullWidth onClick={createRedpacket}>
                     {t('plugin_red_packet_send_symbol', {
                         amount: formatBalance(settings?.total, settings?.token?.decimals ?? 0),
                         symbol: settings?.token?.symbol,
