+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketDialog.tsx
@@ -1,18 +1,10 @@
 import { useCompositionContext } from '@masknet/plugin-infra/content-script'
-import { InjectedDialog, useOpenShareTxDialog } from '@masknet/shared'
+import { InjectedDialog } from '@masknet/shared'
 import { useRemoteControlledDialog } from '@masknet/shared-base-ui'
 import { makeStyles } from '@masknet/theme'
-import {
-    getChainName,
-    TransactionStateType,
-    useAccount,
-    useChainId,
-    useNetworkType,
-    useRedPacketConstants,
-    useWeb3,
-} from '@masknet/web3-shared-evm'
+import { useAccount, useChainId } from '@masknet/web3-shared-evm'
 import { DialogContent } from '@mui/material'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useState } from 'react'
 import Web3Utils from 'web3-utils'
 import {
     useCurrentIdentity,
@@ -24,9 +16,8 @@ import Services from '../../../extension/service'
 import { useI18N } from '../../../utils'
 import { WalletMessages } from '../../Wallet/messages'
 import { RedPacketMetaKey } from '../constants'
-import { RedPacketRPC } from '../messages'
-import { DialogTabs, RedPacketJSONPayload, RedPacketRecord, RpTypeTabs } from '../types'
-import { RedPacketSettings, useCreateCallback } from './hooks/useCreateCallback'
+import { DialogTabs, RedPacketJSONPayload, RpTypeTabs } from '../types'
+import type { RedPacketSettings } from './hooks/useCreateCallback'
 import { RedPacketConfirmDialog } from './RedPacketConfirmDialog'
 import { RedPacketCreateNew } from './RedPacketCreateNew'
 import { RedPacketPast } from './RedPacketPast'
@@ -76,14 +67,10 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
     const chainId = useChainId()
     const account = useAccount()
     const { classes } = useStyles()
-    const { HAPPY_RED_PACKET_ADDRESS_V4 } = useRedPacketConstants()
     const { attachMetadata, dropMetadata } = useCompositionContext()
     const state = useState(DialogTabs.create)
 
-    const networkType = useNetworkType()
-    const contract_version = 4
     const [settings, setSettings] = useState<RedPacketSettings>()
-    const web3 = useWeb3()
 
     const onClose = useCallback(() => {
         setStep(CreateRedPacketPageStep.NewRedPacketPage)
@@ -92,8 +79,6 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
         setValue(DialogTabs.create)
         props.onClose()
     }, [props, state])
-
-    const { address: publicKey, privateKey } = useMemo(() => web3.eth.accounts.create(), [])
 
     const currentIdentity = useCurrentIdentity()
 
@@ -129,98 +114,6 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
         },
         [onClose, chainId, senderName],
     )
-
-    // #region blocking
-    // password should remain the same rather than change each time when createState change,
-    //  otherwise password in database would be different from creating red-packet.
-    const [createSettings, createState, createCallback, resetCreateCallback] = useCreateCallback(
-        settings!,
-        contract_version,
-        publicKey,
-    )
-    const openShareTxDialog = useOpenShareTxDialog()
-    const createRedpacket = useCallback(async () => {
-        const hash = await createCallback()
-        if (hash) {
-            await openShareTxDialog({
-                hash,
-                onShare() {},
-            })
-            // reset state
-            resetCreateCallback()
-
-            // the settings is not available
-            if (!createSettings?.token) return
-
-            // TODO: early return happened
-            // we should guide user to select the red packet in the existing list
-            if (createState.type !== TransactionStateType.CONFIRMED) return
-
-            const { receipt } = createState
-            const CreationSuccess = (receipt.events?.CreationSuccess.returnValues ?? {}) as {
-                creation_time: string
-                creator: string
-                id: string
-                token_address: string
-                total: string
-            }
-            payload.current.sender = {
-                address: account,
-                name: createSettings.name,
-                message: createSettings.message,
-            }
-            payload.current.is_random = createSettings.isRandom
-            payload.current.shares = createSettings.shares
-            payload.current.password = privateKey
-            payload.current.rpid = CreationSuccess.id
-            payload.current.total = CreationSuccess.total
-            payload.current.duration = createSettings.duration
-            payload.current.creation_time = Number.parseInt(CreationSuccess.creation_time, 10) * 1000
-            payload.current.token = createSettings.token
-
-            setSettings(undefined)
-            // output the redpacket as JSON payload
-            onCreateOrSelect(payload.current)
-        }
-    }, [createCallback, resetCreateCallback, createSettings, createState, openShareTxDialog, onCreateOrSelect])
-    // #endregion
-
-    // assemble JSON payload
-    const payload = useRef<RedPacketJSONPayload>({
-        network: getChainName(chainId),
-    } as RedPacketJSONPayload)
-
-    useEffect(() => {
-        if (createState.type !== TransactionStateType.UNKNOWN) return
-        const contractAddress = HAPPY_RED_PACKET_ADDRESS_V4
-        if (!contractAddress) {
-            onClose()
-            return
-        }
-        payload.current.contract_address = contractAddress
-        payload.current.contract_version = contract_version
-        payload.current.network = getChainName(chainId)
-    }, [chainId, networkType, contract_version, createState])
-
-    // open the transaction dialog
-    useEffect(() => {
-        if (!createSettings?.token || createState.type === TransactionStateType.UNKNOWN) return
-
-        // storing the created red packet in DB, it helps retrieve red packet password later
-        // save to the database early, otherwise red-packet would lose when close the tx dialog or
-        //  web page before create successfully.
-        if (createState.type === TransactionStateType.HASH && createState.hash) {
-            payload.current.txid = createState.hash
-            const record: RedPacketRecord = {
-                id: createState.hash!,
-                from: '',
-                password: privateKey,
-                contract_version,
-            }
-            RedPacketRPC.discoverRedPacket(record)
-        }
-    }, [createState /* update tx dialog only if state changed */])
-    // #endregion
 
     const [step, setStep] = useState(CreateRedPacketPageStep.NewRedPacketPage)
     const onBack = useCallback(() => {
@@ -270,8 +163,16 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
         },
     }
 
-    const isCreating = step === CreateRedPacketPageStep.NewRedPacketPage
-    const title = isCreating ? t('plugin_red_packet_display_name') : t('plugin_red_packet_details')
+    const handleCreated = useCallback(
+        (payload: RedPacketJSONPayload) => {
+            onCreateOrSelect(payload)
+            setSettings(undefined)
+        },
+        [onCreateOrSelect],
+    )
+
+    const isCreateStep = step === CreateRedPacketPageStep.NewRedPacketPage
+    const title = isCreateStep ? t('plugin_red_packet_display_name') : t('plugin_red_packet_details')
 
     return (
         <InjectedDialog open={props.open} title={title} onClose={onClose} disableTitleBorder>
@@ -283,7 +184,7 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
                     <RedPacketConfirmDialog
                         onClose={onClose}
                         onBack={onBack}
-                        onCreate={createRedpacket}
+                        onCreated={handleCreated}
                         settings={settings}
                     />
                 ) : null}
